many refactor things-2
