+++ .style.yapf
@@ -87,12 +87,3 @@ SPLIT_BEFORE_DOT = false
 #USE_TABS = true
 
 #SPLIT_PENALTY_LOGICAL_OPERATOR = false
-
-
-
-
-
-
-
-
-
