+++ extramodules/stringOperations.py
@@ -1,0 +1,55 @@
+#!/usr/bin/env python
+# PYTHON_ARGCOMPLETE_OK
+# -*- coding: utf-8 -*-
+
+# Copyright 2019-2020 CERN and copyright holders of ALICE O2.
+# See https://alice-o2.web.cern.ch/copyright for details of the copyright holders.
+# All rights not expressly granted are reserved.
+#
+# This software is distributed under the terms of the GNU General Public
+# License v3 (GPL Version 3), copied verbatim in the file "COPYING".
+#
+# In applying this license CERN does not waive the privileges and immunities
+# granted to it by virtue of its status as an Intergovernmental Organization
+# or submit itself to any jurisdiction.
+
+# This script includes string operations for logic
+
+
+def listToString(s: list):
+    """
+    ListToString provides converts lists to strings with commas.
+    This function is written to save as string type instead of list
+
+
+    Args:
+        s (list): Input as List
+
+    Returns:
+        string: Comma seperated string
+    """
+    if len(s) > 1:
+        # initialize an empty string
+        str1 = ","
+        
+        # return string
+        return str1.join(s)
+    else:
+        str1 = " "
+        
+        return str1.join(s)
+
+
+def stringToList(string: str):
+    """
+    stringToList provides converts strings to list with commas.
+    This function is written to save as list type instead of string
+
+    Args:
+        string (string): Input as String
+
+    Returns:
+        list: merge string elements with comma 
+    """
+    li = list(string.split(","))
+    return li
