+++ runDQEfficiency.py
@@ -22,35 +22,20 @@ import json
 import sys
 import logging
 import logging.config
-from logging import handlers
 import os
-import argparse
+from extramodules.configGetter import configGetter
+from extramodules.debugSettings import debugSettings
 
-from extramodules.actionHandler import NoAction
-from extramodules.actionHandler import ChoicesAction
-from extramodules.helperOptions import HelperOptions
-from extramodules.dqOperations import listToString, stringToList, multiConfigurableSet
-from extramodules.dqOperations import (CfgInvalidFormatError, ForgettedArgsError, NotInAlienvError, runPycacheRemover)
+from extramodules.monitoring import dispArgs
+from extramodules.dqTranscations import aodFileChecker, forgettedArgsChecker, jsonTypeChecker, mainTaskChecker
+from extramodules.configSetter import multiConfigurableSet
+from extramodules.pycacheRemover import runPycacheRemover
 
 from dqtasks.dqEfficiency import DQEfficiency
-"""
-argcomplete - Bash tab completion for argparse
-Documentation https://kislyuk.github.io/argcomplete/
-Instalation Steps
-pip install argcomplete
-sudo activate-global-python-argcomplete
-Only Works On Local not in O2
-Activate libraries in below and activate #argcomplete.autocomplete(parser) line
-"""
-import argcomplete
-from argcomplete.completers import ChoicesCompleter
 
 ###################################
 # Interface Predefined Selections #
 ###################################
-
-readerPath = "configs/readerConfiguration_reducedEventMC.json"
-writerPath = "configs/writerConfiguration_dileptonMC.json"
 
 booleanSelections = ["true", "false"]
 
@@ -58,13 +43,6 @@ isAnalysisEventSelected = True
 isAnalysisTrackSelected = True
 isAnalysisMuonSelected = True
 isAnalysisSameEventPairingSelected = True
-
-O2DPG_ROOT = os.environ.get("O2DPG_ROOT")
-QUALITYCONTROL_ROOT = os.environ.get("QUALITYCONTROL_ROOT")
-O2_ROOT = os.environ.get("O2_ROOT")
-O2PHYSICS_ROOT = os.environ.get("O2PHYSICS_ROOT")
-
-threeSelectedList = []
 
 # List for Selected skimmed process functions for dummy automizer
 skimmedListEventSelection = []
@@ -77,79 +55,8 @@ skimmedListDileptonTrack = []
 # Init Workflow #
 #################
 
-
-class RunDQEfficiency(object):
-    
-    """
-    This class is for managing the workflow by using the interface arguments from
-    all other Common dependencies and the dqEfficiency Task's own arguments in a combined structure.
-
-    Args:
-      object (parser_args() object): runDQEfficiency.py workflow
-    """
-    
-    def __init__(
-            self, parserRunDQEfficiency = argparse.ArgumentParser(
-                formatter_class = argparse.ArgumentDefaultsHelpFormatter,
-                description = "Example Usage: ./runDQEfficiency.py <yourConfig.json> --arg value "
-                ), dqEfficiency = DQEfficiency(), helperOptions = HelperOptions(),
-        ):
-        super(RunDQEfficiency, self).__init__()
-        self.parserRunDQEfficiency = parserRunDQEfficiency
-        self.dqEfficiency = dqEfficiency
-        self.helperOptions = helperOptions
-        self.parserRunDQEfficiency.register("action", "none", NoAction)
-        self.parserRunDQEfficiency.register("action", "store_choice", ChoicesAction)
-    
-    def addArguments(self):
-        """
-        This function allows to add arguments for parser_args() function
-        """
-        
-        # Core Part
-        groupCoreSelections = self.parserRunDQEfficiency.add_argument_group(title = "Core configurations that must be configured")
-        groupCoreSelections.add_argument("cfgFileName", metavar = "Config.json", default = "config.json", help = "config JSON file name",)
-        
-        # aod
-        groupDPLReader = self.parserRunDQEfficiency.add_argument_group(title = "Data processor options: internal-dpl-aod-reader")
-        groupDPLReader.add_argument("--aod", help = "Add your AOD File with path", action = "store", type = str)
-        groupDPLReader.add_argument(
-            "--reader",
-            help = "Reader config JSON with path. For Standart Analysis use as default, for dilepton analysis change to dilepton JSON config file",
-            action = "store", default = readerPath, type = str,
-            )
-        groupDPLReader.add_argument(
-            "--writer", help = "Argument for producing dileptonAOD.root. Set false for disable", action = "store", default = writerPath,
-            type = str,
-            )
-                    
-    def parseArgs(self):
-        """
-        This function allows to save the obtained arguments to the parser_args() function
-
-        Returns:
-            Namespace: returns parse_args()
-        """
-        
-        argcomplete.autocomplete(self.parserRunDQEfficiency, always_complete_options = False)
-        return self.parserRunDQEfficiency.parse_args()
-    
-    def mergeArgs(self):
-        """
-        This function allows to merge parser_args argument information from different classes
-        """
-        
-        self.helperOptions.parserHelperOptions = self.parserRunDQEfficiency
-        self.helperOptions.addArguments()
-        
-        self.dqEfficiency.parserDQEfficiency = self.parserRunDQEfficiency
-        self.dqEfficiency.addArguments()
-        
-        self.addArguments()
-
-
 # init args manually
-initArgs = RunDQEfficiency()
+initArgs = DQEfficiency()
 initArgs.mergeArgs()
 initArgs.parseArgs()
 
@@ -157,109 +64,26 @@ args = initArgs.parseArgs()
 configuredCommands = vars(args) # for get args
 
 # Debug Settings
-if args.debug and (not args.logFile):
-    DEBUG_SELECTION = args.debug
-    numeric_level = getattr(logging, DEBUG_SELECTION.upper(), None)
-    if not isinstance(numeric_level, int):
-        raise ValueError("Invalid log level: %s" % DEBUG_SELECTION)
-    logging.basicConfig(format = "[%(levelname)s] %(message)s", level = DEBUG_SELECTION)
+debugSettings(args.debug, args.logFile, fileName = "dqEfficiency.log")
 
-if args.logFile and args.debug:
-    log = logging.getLogger("")
-    level = logging.getLevelName(args.debug)
-    log.setLevel(level)
-    format = logging.Formatter("%(asctime)s - [%(levelname)s] %(message)s")
-    
-    ch = logging.StreamHandler(sys.stdout)
-    ch.setFormatter(format)
-    log.addHandler(ch)
-    
-    loggerFile = "tableReader.log"
-    if os.path.isfile(loggerFile):
-        os.remove(loggerFile)
-    
-    fh = handlers.RotatingFileHandler(loggerFile, maxBytes = (1048576 * 5), backupCount = 7, mode = "w")
-    fh.setFormatter(format)
-    log.addHandler(fh)
-
-# Transcation management for forgettining assign a value to parameters
-forgetParams = []
-for key, value in configuredCommands.items():
-    if value is not None:
-        if (isinstance(value, str) or isinstance(value, list)) and len(value) == 0:
-            forgetParams.append(key)
-try:
-    if len(forgetParams) > 0:
-        raise ForgettedArgsError(forgetParams)
-except ForgettedArgsError as e:
-    logging.exception(e)
-    sys.exit()
+# Transcation management
+forgettedArgsChecker(configuredCommands)
 
 # Get Some cfg values provided from --param
-for keyCfg, valueCfg in configuredCommands.items():
-    if valueCfg is not None: # Skipped None types, because can"t iterate in None type
-        if keyCfg == "analysis":
-            if isinstance(valueCfg, str):
-                valueCfg = stringToList(valueCfg)
-            analysisCfg = valueCfg
-        if keyCfg == "process":
-            if isinstance(valueCfg, str):
-                valueCfg = stringToList(valueCfg)
-            processCfg = valueCfg
-# Make some checks on provided arguments
-# if len(sys.argv) < 2:
-# logging.error("Invalid syntax! The command line should look like this:")
-# logging.info(" ./runDQEfficiency.py <yourConfig.json> --param value ...")
-# sys.exit()
+analysisCfg = configGetter(configuredCommands, "analysis")
+processCfg = configGetter(configuredCommands, "process")
 
 # Load the configuration file provided as the first parameter
-cfgControl = sys.argv[1] == args.cfgFileName
-isConfigJson = sys.argv[1].endswith(".json")
 config = {}
-
-try:
-    if cfgControl:
-        if not isConfigJson:
-            raise CfgInvalidFormatError(sys.argv[1])
-        else:
-            logging.info("%s is valid json config file", args.cfgFileName)
-
-except CfgInvalidFormatError as e:
-    logging.exception(e)
-    sys.exit()
-
-with open(sys.argv[1]) as configFile:
+with open(args.cfgFileName) as configFile:
     config = json.load(configFile)
-"""
-try:
-    if cfgControl:
-        with open(args.cfgFileName) as configFile:
-            config = json.load(configFile)
-    else:
-        logging.error("Invalid syntax! After the script you must define your json configuration file!!! The command line should look like this:")
-        logging.info("  ./runDQEfficiency.py <yourConfig.json> <-runData|-runMC> --param value ...")
-        sys.exit()
-except FileNotFoundError:
-    isConfigJson = sys.argv[1].endswith(".json")
-    if not isConfigJson:
-            logging.error("Invalid syntax! After the script you must define your json configuration file!!! The command line should look like this:")
-            logging.info(" ./runDQEfficiency.py <yourConfig.json> --param value ...")
-            sys.exit()
-    logging.error("Your JSON Config File found in path!!!")
-    sys.exit()
-"""
+
+jsonTypeChecker(args.cfgFileName)
 
 taskNameInCommandLine = "o2-analysis-dq-efficiency"
+taskNameInConfig = "analysis-event-selection"
 
-# Check alienv
-try:
-    if O2PHYSICS_ROOT is None:
-        raise NotInAlienvError
-    else:
-        logging.info("You are in %s alienv", O2PHYSICS_ROOT)
-except NotInAlienvError as e:
-    logging.exception(e)
-    sys.exit()
+mainTaskChecker(config, taskNameInConfig)
 
 #############################
 # Start Interface Processes #
@@ -513,40 +337,7 @@ for key, value in config.items():
                     else:
                         config[key]["processDummy"] = "true"
 
-# AOD File and Reader-Writer Checker
-if args.aod is not None:
-    argProvidedAod = args.aod
-    textAodList = argProvidedAod.startswith("@")
-    endsWithRoot = argProvidedAod.endswith(".root")
-    endsWithTxt = argProvidedAod.endswith("txt") or argProvidedAod.endswith("text")
-    if textAodList and endsWithTxt:
-        argProvidedAod = argProvidedAod.replace("@", "")
-        logging.info("You provided AO2D list as text file : %s", argProvidedAod)
-        try:
-            open(argProvidedAod, "r")
-            logging.info("%s has valid File Format and Path, File Found", argProvidedAod)
-        
-        except FileNotFoundError:
-            logging.exception("%s AO2D file text list not found in path!!!", argProvidedAod)
-            sys.exit()
-    
-    elif endsWithRoot:
-        logging.info("You provided single AO2D root file : %s", argProvidedAod)
-        try:
-            open(argProvidedAod, "r")
-            logging.info("%s has valid File Format and Path, File Found", argProvidedAod)
-        
-        except FileNotFoundError:
-            logging.exception("%s AO2D single root file not found in path!!!", argProvidedAod)
-            sys.exit()
-    else:
-        try:
-            open(argProvidedAod, "r")
-            logging.info("%s has valid File Format and Path, File Found", argProvidedAod)
-        
-        except FileNotFoundError:
-            logging.exception("%s Wrong formatted File, check your file extension!", argProvidedAod)
-            sys.exit()
+aodFileChecker(args.aod)
 
 if args.reader is not None:
     if not os.path.isfile(args.reader):
@@ -576,13 +367,7 @@ logging.info(commandToRun)
 print("====================================================================================================================")
 
 # Listing Added Commands
-logging.info("Args provided configurations List")
-print("====================================================================================================================")
-for key, value in configuredCommands.items():
-    if value is not None:
-        if isinstance(value, list):
-            listToString(value)
-        logging.info("--%s : %s ", key, value)
+dispArgs(configuredCommands)
 
 os.system(commandToRun)
