+++ runFilterPP.py
@@ -19,18 +19,18 @@
 # Orginal Task: https://github.com/AliceO2Group/O2Physics/blob/master/PWGDQ/Tasks/filterPP.cxx
 
 import json
-import sys
 import logging
 import logging.config
-from logging import handlers
 import os
+from extramodules.debugSettings import debugSettings
 
-from extramodules.dqOperations import listToString, multiConfigurableSet
-from extramodules.dqOperations import (
-    runPycacheRemover, aodFileChecker, filterSelsTranscation, trackPropTransaction, forgettedArgsChecker, jsonTypeChecker, mainTaskChecker
-    )
+from extramodules.monitoring import dispArgs
+from extramodules.dqTranscations import aodFileChecker, forgettedArgsChecker, jsonTypeChecker, filterSelsTranscation, mainTaskChecker, trackPropTransaction
+from extramodules.configSetter import multiConfigurableSet
+from extramodules.pycacheRemover import runPycacheRemover
 
 from dqtasks.filterPP import DQFilterPPTask
+# from extramodules.getTTrees import getTTrees # activate when we have no performance issue
 
 ###################################
 # Interface Predefined Selections #
@@ -64,30 +64,7 @@ args = initArgs.parseArgs()
 configuredCommands = vars(args) # for get args
 
 # Debug Settings
-if args.debug and (not args.logFile):
-    DEBUG_SELECTION = args.debug
-    numeric_level = getattr(logging, DEBUG_SELECTION.upper(), None)
-    if not isinstance(numeric_level, int):
-        raise ValueError("Invalid log level: %s" % DEBUG_SELECTION)
-    logging.basicConfig(format = "[%(levelname)s] %(message)s", level = DEBUG_SELECTION)
-
-if args.logFile and args.debug:
-    log = logging.getLogger("")
-    level = logging.getLevelName(args.debug)
-    log.setLevel(level)
-    format = logging.Formatter("%(asctime)s - [%(levelname)s] %(message)s")
-    
-    ch = logging.StreamHandler(sys.stdout)
-    ch.setFormatter(format)
-    log.addHandler(ch)
-    
-    loggerFile = "filterPP.log"
-    if os.path.isfile(loggerFile):
-        os.remove(loggerFile)
-    
-    fh = handlers.RotatingFileHandler(loggerFile, maxBytes = (1048576 * 5), backupCount = 7, mode = "w")
-    fh.setFormatter(format)
-    log.addHandler(fh)
+debugSettings(args.debug, args.logFile, fileName = "filterPP.log")
 
 forgettedArgsChecker(configuredCommands)
 
@@ -327,6 +304,16 @@ for key, value in config.items():
 filterSelsTranscation(args.cfgBarrelSels, args.cfgMuonSels, args.cfgBarrelTrackCuts, args.cfgMuonsCuts, configuredCommands)
 aodFileChecker(args.aod)
 trackPropTransaction(args.add_track_prop, commonDeps)
+"""
+# Converter Management
+if args.aod is not None:
+    ttreeList = getTTrees(args.aod)
+else:
+    ttreeList = config["internal-dpl-aod-reader"]["aod-file"]
+
+converterManager(ttreeList, commonDeps)
+trackPropChecker(commonDeps, commonDeps)
+"""
 
 ###########################
 # End Interface Processes #
@@ -368,13 +355,7 @@ logging.info(commandToRun)
 print("====================================================================================================================")
 
 # Listing Added Commands
-logging.info("Args provided configurations List")
-print("====================================================================================================================")
-for key, value in configuredCommands.items():
-    if value is not None:
-        if isinstance(value, list):
-            listToString(value)
-        logging.info("--%s : %s ", key, value)
+dispArgs(configuredCommands)
 
 os.system(commandToRun)
