+++ runTableMakerMC.py
@@ -19,16 +19,16 @@
 # Orginal Task: https://github.com/AliceO2Group/O2Physics/blob/master/PWGDQ/TableProducer/tableMakerMC.cxx
 
 import json
-import sys
 import logging
 import logging.config
-from logging import handlers
 import os
 
-from extramodules.dqOperations import listToString, multiConfigurableSet
-from extramodules.dqOperations import (
-    runPycacheRemover, forgettedArgsChecker, jsonTypeChecker, mainTaskChecker, centTranscation, aodFileChecker, trackPropTransaction
-    )
+from extramodules.debugSettings import debugSettings
+from extramodules.monitoring import dispArgs
+from extramodules.descriptor import inputDescriptors, outputDescriptors
+from extramodules.dqTranscations import aodFileChecker, centTranscation, forgettedArgsChecker, jsonTypeChecker, mainTaskChecker, trackPropTransaction
+from extramodules.configSetter import multiConfigurableSet
+from extramodules.pycacheRemover import runPycacheRemover
 
 from dqtasks.tableMakerMC import TableMakerMC
 
@@ -185,30 +185,7 @@ args = initArgs.parseArgs()
 configuredCommands = vars(args) # for get args
 
 # Debug Settings
-if args.debug and (not args.logFile):
-    DEBUG_SELECTION = args.debug
-    numeric_level = getattr(logging, DEBUG_SELECTION.upper(), None)
-    if not isinstance(numeric_level, int):
-        raise ValueError("Invalid log level: %s" % DEBUG_SELECTION)
-    logging.basicConfig(format = "[%(levelname)s] %(message)s", level = DEBUG_SELECTION)
-
-if args.logFile and args.debug:
-    log = logging.getLogger("")
-    level = logging.getLevelName(args.debug)
-    log.setLevel(level)
-    format = logging.Formatter("%(asctime)s - [%(levelname)s] %(message)s")
-    
-    ch = logging.StreamHandler(sys.stdout)
-    ch.setFormatter(format)
-    log.addHandler(ch)
-    
-    loggerFile = "tableMakerMC.log"
-    if os.path.isfile(loggerFile):
-        os.remove(loggerFile)
-    
-    fh = handlers.RotatingFileHandler(loggerFile, maxBytes = (1048576 * 5), backupCount = 7, mode = "w")
-    fh.setFormatter(format)
-    log.addHandler(fh)
+debugSettings(args.debug, args.logFile, fileName = "tableMakerMC.log")
 
 # Transcation management
 forgettedArgsChecker(configuredCommands)
@@ -432,6 +409,16 @@ for key, value in config.items():
 centTranscation(config, args.process, args.syst, centSearch)
 aodFileChecker(args.aod)
 trackPropTransaction(args.add_track_prop, barrelDeps)
+"""
+# Converter Management
+if args.aod is not None:
+    ttreeList = getTTrees(args.aod)
+else:
+    ttreeList = config["internal-dpl-aod-reader"]["aod-file"]
+
+converterManager(ttreeList, commonDeps)
+trackPropChecker(commonDeps, barrelDeps)
+"""
 
 ###########################
 # End Interface Processes #
@@ -494,42 +481,12 @@ for processFunc in specificDeps.keys():
             logging.info("%s", table)
             tablesToProduce[table] = 1
 
-# Generate the aod-writer output descriptor json file
-writerConfig = {}
-writerConfig["OutputDirector"] = {
-    "debugmode": True,
-    "resfile": "reducedAod",
-    "resfilemode": "RECREATE",
-    "ntfmerge": 1,
-    "OutputDescriptors": [],
-    }
-
-# Generate the aod-reader output descriptor json file
-readerConfig = {}
-readerConfig["InputDirector"] = {
-    "debugmode": True,
-    "InputDescriptors": []
-    }
-
-iTable = 0
-for table in tablesToProduce.keys():
-    writerConfig["OutputDirector"]["OutputDescriptors"].insert(iTable, tables[table])
-    readerConfig["InputDirector"]["InputDescriptors"].insert(iTable, tables[table])
-    iTable += 1
-
-writerConfigFileName = "aodWriterTempConfig.json"
-with open(writerConfigFileName, "w") as writerConfigFile:
-    json.dump(writerConfig, writerConfigFile, indent = 2)
-
 readerConfigFileName = "aodReaderTempConfig.json"
-with open(readerConfigFileName, "w") as readerConfigFile:
-    json.dump(readerConfig, readerConfigFile, indent = 2)
+writerConfigFileName = "aodWriterTempConfig.json"
 
-logging.info("aodWriterTempConfig==========")
-print(writerConfig)
-# sys.exit()
-logging.info("aodReaderTempConfig==========")
-print(readerConfig)
+# Generate the aod-writer output descriptor json file
+outputDescriptors(tablesToProduce, tables)
+inputDescriptors(tablesToProduce, tables)
 
 commandToRun = (
     taskNameInCommandLine + " --configuration json://" + updatedConfigFileName +
@@ -568,13 +525,7 @@ print("=========================================================================
 # sys.exit()
 
 # Listing Added Commands
-logging.info("Args provided configurations List")
-print("====================================================================================================================")
-for key, value in configuredCommands.items():
-    if value is not None:
-        if isinstance(value, list):
-            listToString(value)
-        logging.info("--%s : %s ", key, value)
+dispArgs(configuredCommands)
 
 os.system(commandToRun)
