+++ runV0selector.py
@@ -24,13 +24,15 @@ import logging
 import logging.config
 from logging import handlers
 import os
+from extramodules.debugSettings import debugSettings
+#from extramodules.converterManager import converterManager
 
-from extramodules.dqOperations import listToString
-from extramodules.dqOperations import (
-    forgettedArgsChecker, runPycacheRemover, aodFileChecker, trackPropTransaction, jsonTypeChecker, mainTaskChecker
-    )
+from extramodules.monitoring import dispArgs
+from extramodules.dqTranscations import aodFileChecker, forgettedArgsChecker, jsonTypeChecker, mainTaskChecker, trackPropChecker, trackPropTransaction
+from extramodules.pycacheRemover import runPycacheRemover
 
 from dqtasks.v0selector import V0selector
+# from extramodules.getTTrees import getTTrees # activate when we have no performance issue
 
 ###################################
 # Interface Predefined Selections #
@@ -64,30 +66,7 @@ args = initArgs.parseArgs()
 configuredCommands = vars(args) # for get args
 
 # Debug Settings
-if args.debug and (not args.logFile):
-    DEBUG_SELECTION = args.debug
-    numeric_level = getattr(logging, DEBUG_SELECTION.upper(), None)
-    if not isinstance(numeric_level, int):
-        raise ValueError("Invalid log level: %s" % DEBUG_SELECTION)
-    logging.basicConfig(format = "[%(levelname)s] %(message)s", level = DEBUG_SELECTION)
-
-if args.logFile and args.debug:
-    log = logging.getLogger("")
-    level = logging.getLevelName(args.debug)
-    log.setLevel(level)
-    format = logging.Formatter("%(asctime)s - [%(levelname)s] %(message)s")
-    
-    ch = logging.StreamHandler(sys.stdout)
-    ch.setFormatter(format)
-    log.addHandler(ch)
-    
-    loggerFile = "v0selector.log"
-    if os.path.isfile(loggerFile):
-        os.remove(loggerFile)
-    
-    fh = handlers.RotatingFileHandler(loggerFile, maxBytes = (1048576 * 5), backupCount = 7, mode = "w")
-    fh.setFormatter(format)
-    log.addHandler(fh)
+debugSettings(args.debug, args.Logfile, fileName = "v0selector.log")
 
 forgettedArgsChecker(configuredCommands)
 
@@ -253,6 +232,15 @@ for key, value in config.items():
 
 aodFileChecker(args.aod)
 trackPropTransaction(args.add_track_prop, commonDeps)
+"""
+if args.aod is not None:
+    ttreeList = getTTrees(args.aod)
+else:
+    ttreeList = config["internal-dpl-aod-reader"]["aod-file"]
+
+#converterManager(ttreeList, commonDeps)
+trackPropChecker(commonDeps, commonDeps)
+"""
 
 ###########################
 # End Interface Processes #
@@ -298,13 +286,7 @@ logging.info(commandToRun)
 print("====================================================================================================================")
 
 # Listing Added Commands
-logging.info("Args provided configurations List")
-print("====================================================================================================================")
-for key, value in configuredCommands.items():
-    if value is not None:
-        if isinstance(value, list):
-            listToString(value)
-        logging.info("--%s : %s ", key, value)
+dispArgs(configuredCommands)
 
 os.system(commandToRun)
